MAKE-659: support pre-signal trigger for clean mongod shutdown on Windows

+++ tracker_windows.go
@@ -2,10 +2,12 @@ package jasper
 
 import (
 	"github.com/mongodb/grip"
+	"github.com/pkg/errors"
 )
 
 type windowsProcessTracker struct {
-	job *Job
+	job        *Job
+	terminated bool
 }
 
 func newProcessTracker(name string) (processTracker, error) {
@@ -17,12 +19,19 @@ func newProcessTracker(name string) (processTracker, error) {
 }
 
 func (t *windowsProcessTracker) add(pid uint) error {
+	if t.job == nil {
+		return errors.New("cannot add process because job is invalid")
+	}
 	return t.job.AssignProcess(pid)
 }
 
 func (t *windowsProcessTracker) cleanup() error {
+	if t.job == nil {
+		return errors.New("cannot close because job is invalid")
+	}
 	catcher := grip.NewBasicCatcher()
 	catcher.Add(t.job.Terminate(0))
 	catcher.Add(t.job.Close())
+	t.job = nil
 	return catcher.Resolve()
 }
